requirements.txt
Django==1.10.4
Pillow==3.4.2
pyserial==3.2.1
SimpleCV==1.3
gunicorn
whitenoise